70% CHEF DE SERVICE screen regle

+++ src/main/java/Gestion_mission_backend/demo/dto/EtapeDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class EtapeDTO {
     private String villeCode;
-    private Integer ordreEtape;
+    private Long ordreEtape;
 }
